+++ server/internal/logic/task/taskdetaillogic.go
@@ -64,6 +64,10 @@ func (l *TaskDetailLogic) TaskDetail(req *types.TaskDetailReq) (resp *types.Task
 		Priority:     task.Priority,
 		CategoryId:   categoryId,
 		CategoryName: categoryName,
+		StartTime:    formatNullTime(task.StartTime),
+		EndTime:      formatNullTime(task.EndTime),
+		Reminder:     formatNullTime(task.Reminder),
+		Tags:         task.Tags,
 		CreateTime:   task.CreateTime.Format("2006-01-02 15:04"),
 		UpdateTime:   task.UpdateTime.Format("2006-01-02 15:04"),
 	}, nil